+++ biengual-core/src/main/java/com/biengual/core/domain/entity/pointhistory/PointHistoryEntity.java
@@ -32,11 +32,11 @@ public class PointHistoryEntity extends BaseEntity {
     @Column(name = "user_id", nullable = false, columnDefinition = "bigint")
     private Long userId;
 
-    @Column(name = "point_change", nullable = false, columnDefinition = "int")
-    private int pointChange;
+    @Column(name = "point_change", nullable = false, columnDefinition = "bigint")
+    private Long pointChange;
 
-    @Column(name = "point_balance", nullable = false, columnDefinition = "int")
-    private int pointBalance;
+    @Column(name = "point_balance", nullable = false, columnDefinition = "bigint")
+    private Long pointBalance;
 
     @Enumerated(EnumType.STRING)
     private PointReason reason;
@@ -46,7 +46,7 @@ public class PointHistoryEntity extends BaseEntity {
 
     @Builder
     private PointHistoryEntity(
-        Long userId, int pointChange, int pointBalance, PointReason reason
+        Long userId, Long pointChange, Long pointBalance, PointReason reason
     ) {
         this.userId = userId;
         this.pointChange = pointChange;
@@ -56,12 +56,12 @@ public class PointHistoryEntity extends BaseEntity {
     }
 
     public static PointHistoryEntity createPointHistory(
-        Long userId, int pointChange, Long currentBalance, PointReason reason
+        Long userId, Long pointChange, Long currentBalance, PointReason reason
     ) {
         return PointHistoryEntity.builder()
             .userId(userId)
             .pointChange(pointChange)
-            .pointBalance(currentBalance.intValue())
+            .pointBalance(currentBalance)
             .reason(reason)
             .build();
     }
